Move zones, cloud drives and folder groups under the CloudFS module, update docs and ut
